backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
# REMOVE psycopg2-binary - NOT NEEDED for SQLite!
# psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
alembic==1.12.1
redis==5.0.1
celery==5.3.4
sendgrid==6.10.0
twilio==8.10.0
google-generativeai==0.3.2
python-dotenv==1.0.0
email-validator==2.0.0.post2
phonenumbers==8.13.27
httpx==0.25.2
jinja2==3.1.2
python-dateutil==2.8.2
bcrypt==4.0.1
greenlet==3.0.1
typing-extensions==4.8.0
PyJWT==2.8.0
cryptography==41.0.7
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
email-validator==2.0.0.post2
bcrypt==4.0.1
